+++ src/main/java/ru/naumen/sd40/log/parser/gc/GCParser.java
@@ -7,37 +7,21 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import org.apache.commons.math3.stat.descriptive.DescriptiveStatistics;
+import org.springframework.stereotype.Component;
 import ru.naumen.perfhouse.interfaces.IDataParser;
+import ru.naumen.sd40.log.parser.IData;
 
+
+@Component
 public class GCParser implements IDataParser
 {
-    private DescriptiveStatistics ds = new DescriptiveStatistics();
-
     private Pattern gcExecutionTime = Pattern.compile(".*real=(.*)secs.*");
 
-    public double getCalculatedAvg()
-    {
-        return roundToTwoPlaces(getSafeDouble(ds.getMean()));
-    }
-
-    public long getGcTimes()
-    {
-        return ds.getN();
-    }
-
-    public double getMaxGcTime()
-    {
-        return roundToTwoPlaces(getSafeDouble(ds.getMax()));
-    }
-
-    public boolean isNan()
-    {
-        return getGcTimes() == 0;
-    }
-
-    public void parseLine(String line)
+    public void parseLine(IData data,String line)
     {
         Matcher matcher = gcExecutionTime.matcher(line);
+
+        IGcData ds = (IGcData) data;
 
         if (matcher.find())
         {
