+++ src/main/java/ru/naumen/sd40/log/parser/gc/GcData.java
@@ -1,0 +1,37 @@
+package ru.naumen.sd40.log.parser.gc;
+
+import org.apache.commons.math3.stat.descriptive.DescriptiveStatistics;
+import ru.naumen.sd40.log.parser.IData;
+
+import static ru.naumen.sd40.log.parser.NumberUtils.getSafeDouble;
+import static ru.naumen.sd40.log.parser.NumberUtils.roundToTwoPlaces;
+
+public class GcData implements IGcData {
+
+    private DescriptiveStatistics ds = new DescriptiveStatistics();
+
+    public double getCalculatedAvg()
+    {
+        return roundToTwoPlaces(getSafeDouble(ds.getMean()));
+    }
+
+    public long getGcTimes()
+    {
+        return ds.getN();
+    }
+
+    public double getMaxGcTime()
+    {
+        return roundToTwoPlaces(getSafeDouble(ds.getMax()));
+    }
+
+    public boolean isNan()
+    {
+        return getGcTimes() == 0;
+    }
+
+    public void addValue(Double data){
+        ds.addValue(data);
+    }
+
+}
