+++ src/main/java/ru/naumen/sd40/log/parser/gc/IGcData.java
@@ -1,0 +1,12 @@
+package ru.naumen.sd40.log.parser.gc;
+
+import ru.naumen.sd40.log.parser.IData;
+
+public interface IGcData extends IData {
+
+    double getCalculatedAvg();
+    long getGcTimes();
+    double getMaxGcTime();
+    boolean isNan();
+    void addValue(Double data);
+}
