+++ src/main/java/ru/naumen/sd40/log/parser/sdng/SdngDataParser.java
@@ -1,30 +1,20 @@
 package ru.naumen.sd40.log.parser.sdng;
 
+import org.springframework.stereotype.Component;
 import ru.naumen.perfhouse.interfaces.IDataParser;
 import ru.naumen.sd40.log.parser.ActionDoneParser;
 import ru.naumen.sd40.log.parser.ErrorParser;
+import ru.naumen.sd40.log.parser.IData;
 
+@Component
 public class SdngDataParser implements IDataParser {
 
-    private ActionDoneParser actionsDone;
-    private ErrorParser errors;
-
-    public SdngDataParser() {
-        actionsDone = new ActionDoneParser();
-        errors = new ErrorParser();
-    }
-
     @Override
-    public void parseLine(String line) {
-        errors.parseLine(line);
-        actionsDone.parseLine(line);
+    public void parseLine(IData data, String line) {
+
+        ISdngData sdngData = (ISdngData) data;
+        sdngData.getErrors().parseLine(line);
+        sdngData.getActionsDone().parseLine(line);
     }
 
-    public ActionDoneParser getActionsDone() {
-        return actionsDone;
-    }
-
-    public ErrorParser getErrors() {
-        return errors;
-    }
 }
